Password Hashing Implemented & User MVC Deprecated Due to Redundancy
